18_HR_Resume_Screening/requirements.txt
langchain
streamlit
openai
tiktoken
python-dotenv
unstructured
pinecone-client
pypdf
sentence-transformers